:hammer: refactor: create components folder
