+++ .eslintrc.json
@@ -71,10 +71,10 @@
         "groups": [
           "/^(react|styled-components)/",
           "module",
-          "/^~/atoms/",
-          "/^~/molecules/",
-          "/^~/organismis/",
-          "/^~/templates/",
+          "/^~/components/atoms/",
+          "/^~/components/molecules/",
+          "/^~/components/organisms/",
+          "/^~/components/templates/",
           "/^~/context/",
           ["parent", "sibling", "index"],
           "/^~/(assets|res)/",
