+++ src/components/atoms/Game/styles.ts
