+++ src/components/atoms/SocialIcon/styles.ts
