+++ src/components/molecules/Player/styles.ts
